backend/pyproject.toml
[project]
name = "public-boost-backend"
version = "0.1.0"
description = "Backend for the Public Boost service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<0.122",
    "uvicorn[standard]>=0.30,<0.39",
    "SQLAlchemy>=2.0,<2.1",
    "alembic>=1.13,<1.18",
    "asyncpg>=0.29,<0.31",
    "pydantic>=2.9,<2.13",
    "pydantic-settings>=2.4,<2.13",
    "redis>=5.0,<5.4",
    "python-jose[cryptography]>=3.3,<3.6",
    "passlib[bcrypt]>=1.7,<1.8",
    "email-validator>=2.0,<3.0",
    "slowapi>=0.1.9,<0.2"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6,<0.15",
    "mypy>=1.11,<1.19",
    "pytest>=8.3,<8.5",
    "pytest-asyncio>=0.23,<0.27",
    "httpx>=0.27,<0.29",
    "coverage[toml]>=7.6,<7.12",
    "pytest-cov>=5.0,<5.1",
    "aiosqlite>=0.21,<0.22"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
ignore_missing_imports = true
strict_optional = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app", "app.*", "migrations", "migrations.*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"